Set the connection when all conditions are true
This refactors the connection detection code so that all of the checks
need to be true before the connection is upgraded to `:internet`. This
was the intention of the previous code, but there were a couple ways
that it could have set the status unintentionally. These weren't seen
since it seems like connection setup is pretty deterministic so far.