Watch connection status for power control watchdog
VintageNet's power management relies on regularly being told that
everything is ok. If it's not told for a long time (configurable, but
usually minutes), then it will restart the network interface.

The previous code monitored the signal strength. So long as there was a
signal, everything was good.

It turns out that there's a case where you can be connected to a cell
tower, but all network traffic is disallowed. Pressing buttons on
Twilio's UI can fix this. However, the device needs to reconnect.

This commit moves the watchdog petting logic to the connectivity check
logic that knows whether higher level network is available. This code
can recover from the issue just described.