Report the active APN in the property table
Now that VintageNetQMI supports selecting from a list of APNs it would
be nice to be able to retrieve this information programmatically.

Also, this will allow easier use of debugging and/or information
gathering by a human as we wont have to compare the ICCID with the
configured prefixes to figure out which APN is in use.