Fix infinite retries to connect after a connection has be established
Sometimes the QMI command to start a network connection takes longer to
process than the timeout point for QMI. When this happens a good network
connection is established but VintageNetQMI does not think there is a
good network connection.

This fix handles the `:no_effect` error reason so that VintageNetQMI does
not continue to retry establishing a connection as that return value, per
the specification, means that there already is a good connection
established.