Fix crash when trying to connect w/o an ICCID
If there are problems reading the ICCID off the SIM card, the APN
selection code repeatedly raises an exception.

This isn't good. It would be much better to log an error and raise an
alarm in a more general way.

This updates the code to validate the ICCID before trying to select an
APN. The code in the `with` was really hard to read, so this modifies
function calls to make it easier. The modifications are not intended to
change any semantics.