Handle all ICCID changes and not just the first one
This ensures that if the ICCID being reported is different from the one
that was previously reported (or the default), that it runs through the
try to connect logic. This means that if there are any hiccups that
cause a nil or corrupt ICCID to flow through, that it's not stuck forever.